fix: error for InputRefback in console because of null pkey in case not provided
